backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
scikit-learn==1.6.1
joblib==1.3.2
numpy==1.26.2
